feat: refactor model.py to separate files for each model
