+++ bishop/backend/app/alembic/env.py
@@ -1,9 +1,9 @@
+from sqlmodel import SQLModel
 import os
 from logging.config import fileConfig
 
 from alembic import context
 from sqlalchemy import engine_from_config, pool
-
 # this is the Alembic Config object, which provides
 # access to the values within the .ini file in use.
 config = context.config
@@ -18,8 +18,9 @@ fileConfig(config.config_file_name)
 # target_metadata = mymodel.Base.metadata
 # target_metadata = None
 
-from app.models import SQLModel  # noqa
-from app.core.config import settings # noqa
+import app.models  # noqa
+
+from app.core.config import settings  # noqa
 
 target_metadata = SQLModel.metadata
