+++ bishop/backend/app/api/routes/items.py
@@ -6,7 +6,8 @@ from fastapi import APIRouter, HTTPException
 from sqlmodel import func, select
 
 from app.api.deps import CurrentUser, SessionDep
-from app.models import Item, ItemCreate, ItemPublic, ItemsPublic, ItemUpdate, Message
+from app.models.Item import Item, ItemCreate, ItemPublic, ItemsPublic, ItemUpdate
+from app.models.Mixin import Message
 
 router = APIRouter()
