+++ bishop/backend/app/api/routes/msg.py
@@ -6,12 +6,14 @@ from fastapi import APIRouter, HTTPException
 from sqlmodel import func, select
 
 from app.api.deps import CurrentUser, SessionDep
-from app.models import (
+from app.models.ChatMessage import (
     ChatMessage,
     ChatMessageCreate,
     ChatMessagePublic,
     ChatMessagesPublic,
     ChatMessageUpdate,
+)
+from app.models.Mixin import (
     Message
 )
