+++ bishop/backend/app/models/Item.py
@@ -1,0 +1,39 @@
+import uuid
+
+from sqlmodel import Field, Relationship, SQLModel
+
+from app.models.User import User
+
+
+class ItemBase(SQLModel):
+    title: str = Field(min_length=1, max_length=255)
+    description: str | None = Field(default=None, max_length=255)
+
+
+class ItemCreate(ItemBase):
+    pass
+
+
+class ItemUpdate(ItemBase):
+    title: str | None = Field(
+        default=None, min_length=1, max_length=255)  # type: ignore
+
+
+class Item(ItemBase, table=True):
+    __tablename__ = "item"
+    id: uuid.UUID = Field(default_factory=uuid.uuid4, primary_key=True)
+    title: str = Field(max_length=255)
+    owner_id: uuid.UUID = Field(
+        foreign_key="user.id", nullable=False, ondelete="CASCADE"
+    )
+    owner: User | None = Relationship(back_populates="items")
+
+
+class ItemPublic(ItemBase):
+    id: uuid.UUID
+    owner_id: uuid.UUID
+
+
+class ItemsPublic(SQLModel):
+    data: list[ItemPublic]
+    count: int
