+++ bishop/backend/app/models/Mixin.py
@@ -1,0 +1,5 @@
+from sqlmodel import SQLModel
+
+
+class Message(SQLModel):
+    message: str
