+++ bishop/backend/app/models/Security.py
@@ -1,0 +1,15 @@
+from sqlmodel import Field, SQLModel
+
+
+class Token(SQLModel):
+    access_token: str
+    token_type: str = "bearer"
+
+
+class TokenPayload(SQLModel):
+    sub: str | None = None
+
+
+class NewPassword(SQLModel):
+    token: str
+    new_password: str = Field(min_length=8, max_length=40)
