+++ bishop/backend/app/models/__init__.py
@@ -1,0 +1,45 @@
+from .Mixin import (
+    Message
+)
+from .LLModelResponse import (
+    LLModelResponseBase,
+    LLModelResponseCreate,
+    LLModelResponseUpdate,
+    LLModelResponse,
+    LLModelResponsePublic,
+    LLModelResponsesPublic,
+)
+
+from .ChatMessage import (
+    ChatMessageBase,
+    ChatMessageCreate,
+    ChatMessageUpdate,
+    ChatMessage,
+    ChatMessagePublic,
+    ChatMessagesPublic,
+)
+from .Item import (
+    ItemBase,
+    ItemCreate,
+    ItemUpdate,
+    Item,
+    ItemPublic,
+    ItemsPublic,
+)
+from .User import (
+    UserBase,
+    UserCreate,
+    UserRegister,
+    UserUpdate,
+    UserUpdateMe,
+    UpdatePassword,
+    User,
+    UserPublic,
+)
+
+
+Message.model_rebuild()
+LLModelResponse.model_rebuild()
+ChatMessage.model_rebuild()
+Item.model_rebuild()
+User.model_rebuild()
