+++ bishop/backend/app/tests/conftest.py
@@ -9,7 +9,8 @@ from sqlmodel.ext.asyncio.session import AsyncSession
 from app.core.config import settings
 from app.core.db import async_engine, init_db
 from app.main import app
-from app.models import Item, User
+from app.models.Item import Item
+from app.models.User import User
 from app.tests.utils.user import authentication_token_from_email
 from app.tests.utils.utils import get_superuser_token_headers
