+++ bishop/backend/app/tests/crud/test_user.py
@@ -5,7 +5,7 @@ import pytest_asyncio
 
 from app import crud
 from app.core.security import verify_password
-from app.models import User, UserCreate, UserUpdate
+from app.models.User import User, UserCreate, UserUpdate
 from app.tests.utils.utils import random_email, random_lower_string
